+++ tests/N98/Magento/Command/ListCommandTest.php
@@ -2,7 +2,6 @@
 
 namespace N98\Magento\Command;
 
-use N98\Magento\Command\PHPUnit\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 
 class ListCommandTest extends TestCase
@@ -13,13 +12,13 @@ class ListCommandTest extends TestCase
 
         $commandTester = new CommandTester($command);
         $commandTester->execute(
-            array(
+            [
                 'command' => 'list',
-            )
+            ]
         );
-    
+
         $this->assertContains(
-            sprintf('n98-magerun2 version %s by netz98 new media GmbH', $this->getApplication()->getVersion()),
+            sprintf('n98-magerun2 version %s by netz98 GmbH', $this->getApplication()->getVersion()),
             $commandTester->getDisplay()
         );
     }