+++ tests/N98/Magento/Command/System/Cron/RunCommandTest.php
@@ -2,7 +2,7 @@
 
 namespace N98\Magento\Command\System\Cron;
 
-use N98\Magento\Command\PHPUnit\TestCase;
+use N98\Magento\Command\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 
 class RunCommandTest extends TestCase
@@ -18,7 +18,7 @@ class RunCommandTest extends TestCase
             'command' => $command->getName(),
             'job'     => 'backend_clean_cache',
         ]);
-    
+
         $this->assertContains('Run Magento\Backend\Cron\CleanCache::execute done', $commandTester->getDisplay());
     }
 }