+++ tests/N98/Magento/Command/System/MaintenanceCommandTest.php
@@ -3,7 +3,7 @@
 namespace N98\Magento\Command\System;
 
 use Magento\Framework\App\MaintenanceMode;
-use N98\Magento\Command\PHPUnit\TestCase;
+use N98\Magento\Command\TestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 
 class MaintenanceCommandTest extends TestCase